+++ e2e/next-sandbox/pages/batching.tsx
@@ -106,8 +106,12 @@ function Sandbox() {
       <h2>Storage</h2>
       <table style={styles.dataTable}>
         <tbody>
-          <Row id="numItems" name="Items count" value={liveMap.size} />
-          <Row id="items" name="Items" value={Array.from(liveMap.entries())} />
+          <Row
+            id="numItems"
+            name="Items count"
+            value={Object.keys(liveMap).length}
+          />
+          <Row id="items" name="Items" value={liveMap} />
         </tbody>
       </table>
     </div>